+++ packages/fluentui/docs/src/examples/components/Button/Usage/ButtonExampleWithTooltip.shorthand.steps.ts
@@ -2,7 +2,7 @@ import { Button } from '@fluentui/react-northstar';
 
 const config: ScreenerTestsConfig = {
   themes: ['teams', 'teamsDark', 'teamsHighContrast'],
-  steps: [builder => builder.hover(`.${Button.className}`).snapshot('Shows tooltip')],
+  steps: [builder => builder.hover(`.${Button.deprecated_className}`).snapshot('Shows tooltip')],
 };
 
 export default config;